requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.2
django-allauth==65.1.0
django-appconf==1.0.6
django-compressor==4.5.1
django-cors-headers==4.6.0
django-filter==24.3
django-markdownify==0.9.5
djangorestframework==3.15.2
idna==3.10
markdown2==2.5.1
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
setuptools==75.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
tinycss2==1.4.0
urllib3==2.2.3
webencodings==0.5.1
